chore: build scripts
