+++ packages/code-stack/tsconfig.test.json
@@ -1,18 +1,9 @@
 {
   "$schema": "https://json.schemastore.org/tsconfig",
   "display": "Test",
-  "compilerOptions": {
-    "composite": true,
-    "module": "ESNext",
-    "target": "ES2020"
-  },
+  "extends": "./tsconfig",
   "include": [
-    "./__tests__",
-    "./src"
-  ],
-  "exclude": [
-    "dist",
-    "build",
-    "node_modules"
+    "__tests__",
+    "src"
   ]
 }
