+++ packages/format/package.json
@@ -6,7 +6,7 @@
   "module": "./dist/index.mjs",
   "types": "./dist/index.d.ts",
   "scripts": {
-    "build": "tsc && node ./scripts/build.mjs"
+    "build": "node ../../scripts/build.mjs"
   },
   "devDependencies": {
     "typescript": "^4.6.4",
